0822 엘릭서 바 추가

+++ 0806_임시프로젝트파일_ 하영/GP_subproject_first/UIHpbar.h
@@ -10,8 +10,8 @@ public:
 	void calcRate(int Inhp,int ID,EObject Intype);
 	virtual void Render(Gdiplus::Graphics* MemG);
 
-	int fullHp;
-	int curHp;
+	float fullHp;
+	float curHp;
 	float rate;
 
 	Gdiplus::Rect* barRect;
